refactor: add worker/worker.go
